fielder_ai: rebuild the literal pool and match 36/59 functions
Plain float literals, reverse function order and a static dolsqrtf2
take .rodata to 100%; compound assignments, named .bss statics, loop
forms and statement reordering lift most of the rest. Fixes a real bug
in fielderAIOutfieldPlayAttemptInd (wrong global tested) and retypes
g_FieldingLogic.baseFielderIsOn as s8, which also matches
fielder.c's handleBodyCheck2.

+++ src/game/fielding/fielder.c
@@ -18114,8 +18114,8 @@ void handleBodyCheck2(int fielderIndex) {
         return;
     }
 
-    if ((s8)g_FieldingLogic.baseFielderIsOn >= 0) {
-        base = (s8)g_FieldingLogic.baseFielderIsOn;
+    if (g_FieldingLogic.baseFielderIsOn >= 0) {
+        base = g_FieldingLogic.baseFielderIsOn;
     }
     if (base < 0) {
         return;
